move more metadata functions
